Convert health check to new format
